routes have been split and tests are passing
