+++ refiner/app/api/v1/demo.py
@@ -8,6 +8,7 @@ from fastapi.responses import FileResponse, JSONResponse
 from ...services.refine import refine, validate_message
 from ...services.utils import read_zip
 
+# create a router instance for this file
 router = APIRouter(prefix="/demo")
