+++ refiner/app/core/examples.py
@@ -1,0 +1,5 @@
+from ..services.utils import read_json_from_assets
+
+# ecr endpoint examples
+ECR_REQUEST_EXAMPLES = read_json_from_assets("sample_refine_ecr_request.json")
+ECR_RESPONSE_EXAMPLES = read_json_from_assets("sample_refine_ecr_response.json")
